# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Valora MCP Server API
  description: |
    # Valora MCP Server API Documentation
    
    The Valora MCP Server provides a secure, scalable Memory Container Protocol implementation
    with enterprise-grade features including API key management, integrations, and chat continuity.
    
    ## Features
    
    - **Memory Management**: Store, retrieve, and search memories
    - **Chat Continuity**: Import and export conversations for AI agents
    - **Integration System**: Webhooks and plugins for external services
    - **Security**: API key authentication with permissions and rate limiting
    - **Monitoring**: Health checks and usage statistics
    
    ## Authentication
    
    All API endpoints require authentication using API keys. Include your API key in the
    Authorization header:
    
    ```
    Authorization: Bearer your-api-key
    ```
    
    ## Rate Limiting
    
    API requests are rate-limited per API key:
    - Default: 100 requests per minute, 1000 per hour
    - Configurable per API key
    - Rate limit headers included in responses
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns detailed error messages:
    - `400` - Bad Request (invalid input)
    - `401` - Unauthorized (invalid API key)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found (resource doesn't exist)
    - `429` - Too Many Requests (rate limit exceeded)
    - `500` - Internal Server Error
    
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /memory/{id}:
    get:
      summary: Get Memory
      description: Retrieve a specific memory by ID
      tags:
        - Memory
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Memory ID
          schema:
            type: string
            example: "memory_123456789"
      responses:
        '200':
          description: Memory retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /memory:
    post:
      summary: Create Memory
      description: Create a new memory
      tags:
        - Memory
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  description: Memory content
                  example: "This is a new memory about AI and machine learning"
                tags:
                  type: array
                  items:
                    type: string
                  description: Tags for categorizing the memory
                  example: ["ai", "machine-learning", "important"]
                metadata:
                  type: object
                  description: Additional metadata
                  example:
                    source: "chat-import"
                    confidence: 0.95
      responses:
        '201':
          description: Memory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /memory/search:
    get:
      summary: Search Memories
      description: Search memories by content, tags, or metadata
      tags:
        - Memory
      security:
        - ApiKeyAuth: []
      parameters:
        - name: query
          in: query
          required: true
          description: Search query
          schema:
            type: string
            example: "machine learning"
        - name: tags
          in: query
          description: Filter by tags
          schema:
            type: string
            example: "ai,important"
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memory'
                  total:
                    type: integer
                    example: 25
                  limit:
                    type: integer
                    example: 10
                  offset:
                    type: integer
                    example: 0
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API key for authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid API key"
        code:
          type: string
          description: Error code
          example: "INVALID_API_KEY"
        details:
          type: object
          description: Additional error details
...
    Memory:
      type: object
      properties:
        id:
          type: string
          example: "memory_123456789"
        content:
          type: string
          example: "This is a memory about AI and machine learning"
        tags:
          type: array
          items:
            type: string
          example: ["ai", "machine-learning", "important"]
        metadata:
          type: object
          properties:
            source:
              type: string
              example: "chat-import"
            confidence:
              type: number
              format: float
              example: 0.95
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...